+++ common/src/main/java/dev/boxadactle/coordinatesdisplay/Bindings.java
@@ -10,12 +10,12 @@ import dev.boxadactle.coordinatesdisplay.marking.MarkGui;
 import dev.boxadactle.coordinatesdisplay.position.Position;
 import dev.boxadactle.coordinatesdisplay.registry.DisplayMode;
 import net.minecraft.client.KeyMapping;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.resources.Identifier;
 import org.lwjgl.glfw.GLFW;
 
 public class Bindings {
 
-    public static final KeyMapping.Category CATEGORY = KeyMapping.Category.register(ResourceLocation.fromNamespaceAndPath(CoordinatesDisplay.MOD_ID, "keybinds"));
+    public static final KeyMapping.Category CATEGORY = KeyMapping.Category.register(Identifier.fromNamespaceAndPath(CoordinatesDisplay.MOD_ID, "keybinds"));
     
     public static final KeyMapping hudEnabled = new KeyMapping("key.coordinatesdisplay.hudenabled", GLFW.GLFW_KEY_H, CATEGORY);